+++ src/components/sets/cell-set-utils.js
@@ -223,52 +223,6 @@ export function nodeToLevelDescendantNamePaths(node, level, prevPath, stopEarly 
 }
 
 /**
- * Return whether it makes sense to show a "complement checked sets"
- * button.
- * @param {object} currTree A tree object.
- * @returns {boolean} Does it make sense?
- */
-export function treeHasCheckedSetsToComplement(currTree, checkedPaths, items) {
-  return (
-    currTree
-    && checkedPaths
-    && checkedPaths.length > 0
-    && treeToComplement(currTree, checkedPaths, items).length > 0
-  );
-}
-
-/**
- * Return whether it makes sense to show a "intersect checked sets"
- * button.
- * @param {object} currTree A tree object.
- * @returns {boolean} Does it make sense?
- */
-export function treeHasCheckedSetsToIntersect(currTree, checkedPaths) {
-  return (
-    currTree
-    && checkedPaths
-    && checkedPaths.length > 1
-    && treeToIntersection(currTree, checkedPaths).length > 0
-  );
-}
-
-/**
- * Return whether it makes sense to show a "union checked sets"
- * button.
- * @param {object} currTree A tree object.
- * @returns {boolean} Does it make sense?
- */
-export function treeHasCheckedSetsToUnion(currTree, checkedPaths) {
-  return (
-    currTree
-    && checkedPaths
-    && checkedPaths.length > 1
-    && treeToUnion(currTree, checkedPaths).length > 0
-  );
-}
-
-
-/**
  * Export the tree by clearing tree state and all node states.
  * @param {object} currTree A tree object.
  * @returns {object} Tree object with tree and node state removed.
@@ -458,20 +412,6 @@ export function treeToExpectedCheckedLevel(currTree, checkedPaths) {
   }
   return result;
 }
-
-
-export function treeToCheckedSetOperations(currTree, checkedPaths, items) {
-  const hasCheckedSetsToUnion = treeHasCheckedSetsToUnion(currTree, checkedPaths);
-  const hasCheckedSetsToIntersect = treeHasCheckedSetsToIntersect(currTree, checkedPaths);
-  const hasCheckedSetsToComplement = treeHasCheckedSetsToComplement(currTree, checkedPaths, items);
-
-  return {
-    hasCheckedSetsToUnion,
-    hasCheckedSetsToIntersect,
-    hasCheckedSetsToComplement,
-  };
-}
-
 
 export function treesConflict(cellSets, testCellSets) {
   const paths = [];